Task3

+++ Tyuiu.ZhukovaYA.Sprint5.Task3.V14.Lib/DataService.cs
@@ -7,8 +7,8 @@ namespace Tyuiu.ZhukovaYA.Sprint5.Task3.V14.Lib
         public string SaveToFileTextData(int x)
         {
             string path = Path.Combine(Path.GetTempPath(), "OutPutFileTask3.bin");
-            double res = (4 * Math.Pow(x, 3)) / (Math.Pow(x, 3) - 1);
-            File.WriteAllText(path, Math.Round(res, 3).ToString());
+            double res = (double)(4 * Math.Pow(x, 3)) / (Math.Pow(x, 3) - 1);
+            File.WriteAllText(path, Math.Round((double)res, 3).ToString());
             return path;
         }
     }
